+++ src/core/server/commands/debugVehicle.ts
@@ -30,6 +30,13 @@ ChatController.addCommand(
     handleAddFactionVehicle,
 );
 
+ChatController.addCommand(
+    'addfactionvehicle',
+    '/addfactionvehiclebyid [faction_id] [model]',
+    PERMISSIONS.ADMIN,
+    handleAddFactionVehicleByID,
+);
+
 function handleTemp(player: alt.Player, model: string): void {
     if (!model) {
         playerFuncs.emit.message(player, ChatController.getDescription('vehicle'));
@@ -81,6 +88,52 @@ function handleAdd(player: alt.Player, model: string): void {
 function handleAddFactionVehicle(player: alt.Player, model: string): void {
     if (!player.data || !player.data.faction) {
         playerFuncs.emit.message(player, `Must be in a faction to add that vehicle to the faction.`);
+        return;
+    }
+
+    if (!model) {
+        playerFuncs.emit.message(player, `Vehicle model not specified`);
+        return;
+    }
+
+    if (player.data.isDead) {
+        playerFuncs.emit.message(player, LocaleController.get(LOCALE_KEYS.CANNOT_PERFORM_WHILE_DEAD));
+        return;
+    }
+
+    const fwd = getVectorInFrontOfPlayer(player, 5);
+
+    try {
+        const veh = VehicleFuncs.tempVehicle(player, model, fwd, new alt.Vector3(0, 0, 0));
+
+        VehicleFuncs.add(
+            {
+                owner: player.data.faction.toString(),
+                ownerType: VEHICLE_OWNERSHIP.FACTION,
+                fuel: 100,
+                model,
+                position: veh.pos,
+                rotation: veh.rot,
+            },
+            false,
+        );
+
+        veh.destroy();
+    } catch (err) {
+        console.log(err);
+        playerFuncs.emit.message(player, LocaleController.get(LOCALE_KEYS.INVALID_VEHICLE_MODEL));
+        return;
+    }
+}
+
+function handleAddFactionVehicleByID(player: alt.Player, id: string, model: string): void {
+    if (!player.data || !player.data.faction) {
+        playerFuncs.emit.message(player, `Must be in a faction to add that vehicle to the faction.`);
+        return;
+    }
+
+    if (!id) {
+        playerFuncs.emit.message(player, `Faction ID was not specified.`);
         return;
     }